+++ presentation/src/main/res/layout/fragment_join_two.xml
@@ -6,7 +6,7 @@
     android:id="@+id/frameLayout3"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".join.JoinTwoFragment">
+    tools:context=".login.JoinTwoFragment">
 
     <TextView
         android:id="@+id/tv_join"
